Introduce new 'Sorting' section
- All sorting stuff (sorts, shuffle, reverse) is extracted into Sorting/NSArray+EKSorting.
- All sorting-methods are now in NSMutableArray to reduce a number of copy/mutableCopy garbage. This makes every algorithm much more readable compared to their previous NSArray variants.
