+++ EKAlgorithms/NSArray+EKStuff.m
@@ -137,26 +137,6 @@
     return returnValue;
 }
 
-#pragma mark - Array reverse
-
-- (NSArray *)reversedArray
-{
-    NSMutableArray *reversedArray = [self mutableCopy];
-    
-    NSUInteger count = reversedArray.count;
-    
-    for (int i = 0; i < count / 2; i++) {
-        [reversedArray exchangeObjectAtIndex:i withObjectAtIndex:(count - 1 - i)];
-    }
-    
-    return [reversedArray copy];
-}
-
-- (NSArray *)CocoaImplementationOfReversedArray
-{
-    return [[self reverseObjectEnumerator] allObjects];
-}
-
 #pragma mark - Intersection of two arrays
 
 - (NSArray *)intersectionWithArray:(NSArray *)secondArray
@@ -217,36 +197,6 @@
     }
     
     return [objects copy];
-}
-
-#pragma mark - Check if array is sorted
-
-- (BOOL)isSorted
-{
-    NSUInteger countMinusOne = self.count - 1;
-
-    for (NSUInteger i = 0; i < countMinusOne; i++) {
-        if ([self[i] compare:self[i + 1]] == NSOrderedDescending) {
-            return NO;
-        }
-    }
-
-    return YES;
-}
-
-#pragma mark - Array shuffle
-
-- (NSArray *)shuffledArray
-{
-    NSUInteger i = self.count;
-    NSMutableArray *shuffledArray = [self mutableCopy];
-    
-    while (i) {
-        NSUInteger randomIndex = arc4random_uniform((u_int32_t)i);
-        [shuffledArray exchangeObjectAtIndex:randomIndex withObjectAtIndex:--i];
-    }
-    
-    return [shuffledArray copy];
 }
 
 #pragma mark - Sum of elements
@@ -363,194 +313,6 @@
     }
     
     return indexOfFoundedObject = -1;
-}
-
-#pragma mark - SORT STUFF
-#pragma mark - Bubble sort
-
-- (NSMutableArray *)bubbleSortedArray
-{
-    NSMutableArray *selfCopy = [self mutableCopy];
-    id temp = nil;
-    
-    for (int i = (int)[selfCopy count] - 2; i >= 0; i--) {
-        for (int j = 0; j <= i; j++) {
-            if ([selfCopy[j] integerValue] > [selfCopy[j + 1] integerValue]) {
-                temp = [selfCopy objectAtIndex:j];
-                [selfCopy replaceObjectAtIndex:j withObject:[selfCopy objectAtIndex:j + 1]];
-                [selfCopy replaceObjectAtIndex:j + 1 withObject:temp];
-            }
-        }
-    }
-    
-    return selfCopy;
-}
-
-#pragma mark - Shell sort
-
-- (NSMutableArray *)shellSortedArray
-{
-    NSMutableArray *selfCopy = [self mutableCopy];
-    id temp = nil;
-    
-    for (NSInteger i = [selfCopy count] / 2; i > 0; i = i / 2) {
-        for (NSInteger j = i; j < [selfCopy count]; j++) {
-            for (NSInteger k = j - i; k >= 0; k = k - i) {
-                if ([[selfCopy objectAtIndex:k + 1] floatValue] >= [[selfCopy objectAtIndex:k] floatValue]) {
-                    break;
-                }
-                else {
-                    temp = [selfCopy objectAtIndex:k];
-                    [selfCopy replaceObjectAtIndex:k withObject:[selfCopy objectAtIndex:k + i]];
-                    [selfCopy replaceObjectAtIndex:k + i withObject:temp];
-                }
-            }
-        }
-    }
-    
-    return selfCopy;
-}
-
-#pragma mark - Merge sort stuff
-
-- (NSMutableArray *)mergeSortedArray
-{
-    NSMutableArray *selfCopy = [self mutableCopy];
-    [selfCopy partitionArrayWithMinimalIndex:0 withMaximumIndex:(NSInteger)[selfCopy count] - 1];
-    
-    return selfCopy;
-}
-
-#pragma mark - Quick sort
-
-- (NSMutableArray *)quickSortedArrayWithLeftIndex:(NSInteger)left withRightIndex:(NSInteger)right
-{
-    NSMutableArray *selfCopy = (NSMutableArray *)self;
-    
-    if ([self isKindOfClass:[NSMutableArray class]] == NO) {
-        selfCopy = [self mutableCopy];
-    }
-    
-    NSInteger i = left;
-    NSInteger j = right;
-
-    id pivotalElement = nil;
-
-    pivotalElement = selfCopy[(left + right) / 2];
-    
-    do {
-        while (([selfCopy[i] floatValue] < [pivotalElement floatValue]) && (i < right)) {
-            i++;
-        }
-        while (([pivotalElement floatValue] < [selfCopy[j] floatValue]) && (j > left)) {
-            j--;
-        }
-        
-        if (i <= j) {
-            [selfCopy exchangeObjectAtIndex:i withObjectAtIndex:j];
-
-            i++;
-            j--;
-        }
-    } while (i <= j);
-    
-    if (left < j) {
-        selfCopy = [selfCopy quickSortedArrayWithLeftIndex:left withRightIndex:j];
-    }
-    if (i < right) {
-        selfCopy = [selfCopy quickSortedArrayWithLeftIndex:i withRightIndex:right];
-    }
-    
-    return selfCopy;
-}
-
-#pragma mark - Insertion sort
-
-- (NSMutableArray *)insertionSortedArray
-{
-    NSMutableArray *selfCopy = [self mutableCopy];
-
-    NSUInteger count = self.count;
-
-    for (int i = 1; i < count; i++) {
-        for (int j = i; (j > 0) && ([selfCopy[j - 1] floatValue] > [selfCopy[j] floatValue]); j--) {
-            [selfCopy exchangeObjectAtIndex:j withObjectAtIndex:(j - 1)];
-        }
-    }
-    
-    return selfCopy;
-}
-
-#pragma mark - Selection sort
-
-- (NSArray *)selectionSortedArray
-{
-    NSUInteger count = self.count;
-
-    NSMutableArray *selfCopy = [self mutableCopy];
-    
-    for (NSUInteger i = 0; i < count; i++) {
-        NSUInteger min = i;
-        
-        for (NSUInteger j = i + 1; j < count; j++) {
-            if ([selfCopy[j] floatValue]  < [selfCopy[min] floatValue]) {
-                min = j;
-            }
-        }
-
-        [selfCopy exchangeObjectAtIndex:i withObjectAtIndex:min];
-    }
-    
-    return [selfCopy copy];
-}
-
-#pragma mark - Partial Selection Sort
-
-- (NSArray *)partialSelectionSortedArray:(NSUInteger)K {
-    NSUInteger count = self.count;
-    NSParameterAssert(K <= count);
-
-    NSMutableArray *newArray = [self mutableCopy];
-
-    NSUInteger minIndex;
-    NSUInteger minValue;
-
-    for (NSUInteger i = 0; i < K; i++) {
-        minIndex = i;
-        minValue = [[newArray objectAtIndex:i] unsignedIntegerValue];
-
-        for (NSUInteger j = i + 1; j < count; j++) {
-            NSUInteger el = [[newArray objectAtIndex:j] unsignedIntegerValue];
-
-            if (el < minValue) {
-                minIndex = j;
-                minValue = el;
-            }
-        }
-
-        [newArray exchangeObjectAtIndex:i withObjectAtIndex:minIndex];
-    }
-
-    return [newArray copy];
-}
-
-
-#pragma mark - Heap sort
-
-- (NSMutableArray *)heapSortedArray
-{
-    NSMutableArray *selfCopy = [self mutableCopy];
-    [selfCopy heapifyArrayWithSize:[selfCopy count]];
-    
-    NSInteger end = [selfCopy count] - 1;
-    
-    while (end > 0) {
-        [selfCopy exchangeObjectAtIndex:end withObjectAtIndex:0];
-        [selfCopy siftDownArrayWithStart:0 end:end - 1];
-        end--;
-    }
-    
-    return selfCopy;
 }
